+++ packages/analytics-js-plugins/src/bugsnag/constants.ts
@@ -31,7 +31,6 @@ const APP_STATE_EXCLUDE_KEYS = [
   'instance', // destination instance objects
   'eventBuffer', // pre-load event buffer (may contain PII)
   'traits',
-  'authToken',
 ];
 
 const BUGSNAG_PLUGIN = 'BugsnagPlugin';