+++ packages/analytics-js-service-worker/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.2.1](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-service-worker@3.2.0...@rudderstack/analytics-js-service-worker@3.2.1) (2024-11-12)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.12.1`
 ## [3.2.0](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-service-worker@3.1.17...@rudderstack/analytics-js-service-worker@3.2.0) (2024-11-08)
 
 ### Dependency Updates